Add Authorization for Blog and Registration Edit and Delete

+++ src/app/api/registrasi/[id]/route.ts
@@ -1,8 +1,17 @@
 import { NextResponse } from "next/server";
 import { prisma } from "@/lib/prisma";
+import { cookies } from "next/headers";
 
 export async function PATCH(request: Request, { params }: { params: { id: string } }) {
     try {
+
+        const cookie = await cookies();
+        const session = cookie.get('session')?.value;
+
+        if (!session) {
+            return NextResponse.json({ success: false, message: "Unauthorized" }, { status: 401 });
+        }   
+
         const body = await request.json();
         const id = parseInt(params.id);
         const updatedRegistrasi = await prisma.registrasi.update({
@@ -21,6 +30,13 @@ export async function PATCH(request: Request, { params }: { params: { id: string
 
 export async function DELETE(request: Request, { params }: { params: { id: string } }) {
     try {
+        const cookie = await cookies();
+        const session = cookie.get('session')?.value;
+
+        if (!session) {
+            return NextResponse.json({ success: false, message: "Unauthorized" }, { status: 401 });
+        }
+
         prisma.dataDivisi.deleteMany({
             where: { id_register: parseInt(params.id) }
         });
